Cleanup SQL draft
