+++ src/Microsoft.Health.Dicom.SqlServer/Features/Query/SqlQueryStoreV27.cs
@@ -16,10 +16,16 @@ using Microsoft.Health.SqlServer.Features.Schema.Model;
 using Microsoft.Health.Dicom.Core.Features.Query.Model;
 using System.Linq;
 using Microsoft.Health.SqlServer.Features.Storage;
+using EnsureThat;
+using Microsoft.Data.SqlClient;
+using Microsoft.Health.Dicom.SqlServer.Extensions;
+using Microsoft.Health.SqlServer;
+using System.Text;
+using Microsoft.Health.Dicom.Core.Features.Model;
 
 namespace Microsoft.Health.Dicom.SqlServer.Features.Query;
 
-internal class SqlQueryStoreV27 : SqlQueryStoreV6
+internal class SqlQueryStoreV27 : SqlQueryStoreV4
 {
     public override SchemaVersion Version => SchemaVersion.V27;
 
@@ -28,6 +34,43 @@ internal class SqlQueryStoreV27 : SqlQueryStoreV6
         ILogger<ISqlQueryStore> logger)
         : base(sqlConnectionWrapperFactory, logger)
     {
+    }
+
+    public override async Task<QueryResult> QueryAsync(
+        int partitionKey,
+        QueryExpression query,
+        CancellationToken cancellationToken)
+    {
+        EnsureArg.IsNotNull(query, nameof(query));
+
+        var results = new List<VersionedInstanceIdentifier>(query.EvaluatedLimit);
+
+        using SqlConnectionWrapper sqlConnectionWrapper = await SqlConnectionWrapperFactory.ObtainSqlConnectionWrapperAsync(cancellationToken);
+        using SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateRetrySqlCommand();
+
+        var stringBuilder = new IndentedStringBuilder(new StringBuilder());
+        var sqlQueryGenerator = new SqlQueryGenerator(stringBuilder, query, new SqlQueryParameterManager(sqlCommandWrapper.Parameters), Version, partitionKey);
+
+        sqlCommandWrapper.CommandText = stringBuilder.ToString();
+        sqlCommandWrapper.LogSqlCommand(Logger);
+
+        using SqlDataReader reader = await sqlCommandWrapper.ExecuteReaderAsync(CommandBehavior.SequentialAccess, cancellationToken);
+        while (await reader.ReadAsync(cancellationToken))
+        {
+            (string studyInstanceUid, string seriesInstanceUid, string sopInstanceUid, long watermark) = reader.ReadRow(
+               VLatest.Instance.StudyInstanceUid,
+               VLatest.Instance.SeriesInstanceUid,
+               VLatest.Instance.SopInstanceUid,
+               VLatest.Instance.Watermark);
+
+            results.Add(new VersionedInstanceIdentifier(
+                    studyInstanceUid,
+                    seriesInstanceUid,
+                    sopInstanceUid,
+                    watermark));
+        }
+
+        return new QueryResult(results);
     }
 
     public override async Task<IReadOnlyCollection<StudyResult>> GetStudyResultAsync(
