+++ src/Microsoft.Health.Dicom.SqlServer/Features/Schema/SchemaVersionConstants.cs
@@ -7,10 +7,6 @@ namespace Microsoft.Health.Dicom.SqlServer.Features.Schema;
 
 public static class SchemaVersionConstants
 {
-    public const int Min = (int)SchemaVersion.V34;
+    public const int Min = (int)SchemaVersion.V38;
     public const int Max = (int)SchemaVersion.V43;
-    public const int SupportExtendedQueryTagSchemaVersion = (int)SchemaVersion.V4;
-    public const int SupportDTAndTMInExtendedQueryTagSchemaVersion = (int)SchemaVersion.V5;
-    public const int SupportDataPartitionSchemaVersion = (int)SchemaVersion.V6;
-    public const int SupportUpsRsSchemaVersion = (int)SchemaVersion.V9;
 }
