--- src/Microsoft.Health.Dicom.SqlServer/Features/Store/SqlIndexDataStoreV23.cs
@@ -1,84 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using EnsureThat;
-using FellowOakDicom;
-using Microsoft.Data.SqlClient;
-using Microsoft.Health.Dicom.Core.Exceptions;
-using Microsoft.Health.Dicom.Core.Features.Common;
-using Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag;
-using Microsoft.Health.Dicom.Core.Models;
-using Microsoft.Health.Dicom.SqlServer.Features.ExtendedQueryTag;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema.Model;
-using Microsoft.Health.SqlServer.Features.Client;
-using Microsoft.Health.SqlServer.Features.Storage;
-
-namespace Microsoft.Health.Dicom.SqlServer.Features.Store;
-
-/// <summary>
-/// Sql IndexDataStore version 23.
-/// </summary>
-internal class SqlIndexDataStoreV23 : SqlIndexDataStoreV10
-{
-    public SqlIndexDataStoreV23(SqlConnectionWrapperFactory sqlConnectionWrapperFactory)
-        : base(sqlConnectionWrapperFactory)
-    {
-    }
-
-    public override SchemaVersion Version => SchemaVersion.V22;
-
-    public override async Task EndCreateInstanceIndexAsync(
-        int partitionKey,
-        DicomDataset dicomDataset,
-        long watermark,
-        IEnumerable<QueryTag> queryTags,
-        FileProperties fileProperties,
-        bool allowExpiredTags,
-        bool hasFrameMetadata,
-        CancellationToken cancellationToken)
-    {
-        EnsureArg.IsNotNull(dicomDataset, nameof(dicomDataset));
-        EnsureArg.IsNotNull(queryTags, nameof(queryTags));
-
-        if (fileProperties != null)
-        {
-            // if we are passing in fileProperties, it means we're using a new binary, but with an old schema
-            throw new BadRequestException(DicomSqlServerResource.SchemaVersionNeedsToBeUpgraded);
-        }
-
-        using (SqlConnectionWrapper sqlConnectionWrapper = await SqlConnectionWrapperFactory.ObtainSqlConnectionWrapperAsync(cancellationToken))
-        using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateRetrySqlCommand())
-        {
-            VLatest.UpdateInstanceStatusV6.PopulateCommand(
-                sqlCommandWrapper,
-                partitionKey,
-                dicomDataset.GetSingleValueOrDefault(DicomTag.StudyInstanceUID, string.Empty),
-                dicomDataset.GetSingleValueOrDefault(DicomTag.SeriesInstanceUID, string.Empty),
-                dicomDataset.GetSingleValueOrDefault(DicomTag.SOPInstanceUID, string.Empty),
-                watermark,
-                (byte)IndexStatus.Created,
-                allowExpiredTags ? null : ExtendedQueryTagDataRowsBuilder.GetMaxTagKey(queryTags),
-                hasFrameMetadata);
-
-            try
-            {
-                await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken);
-            }
-            catch (SqlException ex)
-            {
-                throw ex.Number switch
-                {
-                    SqlErrorCodes.NotFound => new InstanceNotFoundException(),
-                    SqlErrorCodes.Conflict when ex.State == 10 => new ExtendedQueryTagsOutOfDateException(),
-                    _ => new DataStoreException(ex),
-                };
-            }
-        }
-    }
-}
