--- src/Microsoft.Health.Dicom.SqlServer/Features/Store/SqlIndexDataStoreV32.cs
@@ -1,134 +1,0 @@
-// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using System.Collections.Generic;
-using System.Data;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using EnsureThat;
-using FellowOakDicom;
-using Microsoft.Data.SqlClient;
-using Microsoft.Health.Dicom.Core.Exceptions;
-using Microsoft.Health.Dicom.Core.Extensions;
-using Microsoft.Health.Dicom.Core.Features.Model;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema.Model;
-using Microsoft.Health.SqlServer.Features.Client;
-using Microsoft.Health.SqlServer.Features.Storage;
-
-namespace Microsoft.Health.Dicom.SqlServer.Features.Store;
-
-/// <summary>
-/// Sql IndexDataStore version 32.
-/// </summary>
-internal class SqlIndexDataStoreV32 : SqlIndexDataStoreV23
-{
-    public SqlIndexDataStoreV32(SqlConnectionWrapperFactory sqlConnectionWrapperFactory)
-        : base(sqlConnectionWrapperFactory)
-    {
-    }
-
-    public override SchemaVersion Version => SchemaVersion.V32;
-    public override async Task<IEnumerable<InstanceMetadata>> BeginUpdateInstanceAsync(int partitionKey, IReadOnlyCollection<long> versions, CancellationToken cancellationToken = default)
-    {
-        var results = new List<InstanceMetadata>();
-
-        using (SqlConnectionWrapper sqlConnectionWrapper = await SqlConnectionWrapperFactory.ObtainSqlConnectionWrapperAsync(cancellationToken))
-        using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateRetrySqlCommand())
-        {
-            List<WatermarkTableTypeRow> versionRows = versions.Select(i => new WatermarkTableTypeRow(i)).ToList();
-            VLatest.BeginUpdateInstance.PopulateCommand(
-                sqlCommandWrapper,
-                partitionKey,
-                versionRows);
-
-            try
-            {
-                using (var reader = await sqlCommandWrapper.ExecuteReaderAsync(CommandBehavior.SequentialAccess, cancellationToken))
-                {
-                    while (await reader.ReadAsync(cancellationToken))
-                    {
-                        (string rStudyInstanceUid,
-                            string rSeriesInstanceUid,
-                            string rSopInstanceUid,
-                            long watermark,
-                            string rTransferSyntaxUid,
-                            bool rHasFrameMetadata,
-                            long? originalWatermark,
-                            long? newWatermark) = reader.ReadRow(
-                               VLatest.Instance.StudyInstanceUid,
-                               VLatest.Instance.SeriesInstanceUid,
-                               VLatest.Instance.SopInstanceUid,
-                               VLatest.Instance.Watermark,
-                               VLatest.Instance.TransferSyntaxUid,
-                               VLatest.Instance.HasFrameMetadata,
-                               VLatest.Instance.OriginalWatermark,
-                               VLatest.Instance.NewWatermark);
-
-                        results.Add(
-                            new InstanceMetadata(
-                                new VersionedInstanceIdentifier(
-                                    rStudyInstanceUid,
-                                    rSeriesInstanceUid,
-                                    rSopInstanceUid,
-                                    watermark,
-                                    partitionKey),
-                                new InstanceProperties()
-                                {
-                                    TransferSyntaxUid = rTransferSyntaxUid,
-                                    HasFrameMetadata = rHasFrameMetadata,
-                                    OriginalVersion = originalWatermark,
-                                    NewVersion = newWatermark
-                                }));
-                    }
-                }
-                return results;
-            }
-            catch (SqlException ex)
-            {
-                throw ex.Number switch
-                {
-                    SqlErrorCodes.NotFound => new InstanceNotFoundException(),
-                    _ => new DataStoreException(ex),
-                };
-            }
-        }
-    }
-
-    public override async Task EndUpdateInstanceAsync(
-        int partitionKey,
-        string studyInstanceUid,
-        DicomDataset dicomDataset,
-        CancellationToken cancellationToken)
-    {
-        EnsureArg.IsNotNull(dicomDataset, nameof(dicomDataset));
-
-        using (SqlConnectionWrapper sqlConnectionWrapper = await SqlConnectionWrapperFactory.ObtainSqlConnectionWrapperAsync(cancellationToken))
-        using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateRetrySqlCommand())
-        {
-            VLatest.EndUpdateInstance.PopulateCommand(
-                sqlCommandWrapper,
-                partitionKey,
-                studyInstanceUid,
-                dicomDataset.GetFirstValueOrDefault<string>(DicomTag.PatientID),
-                dicomDataset.GetFirstValueOrDefault<string>(DicomTag.PatientName),
-                dicomDataset.GetStringDateAsDate(DicomTag.PatientBirthDate));
-
-            try
-            {
-                await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken);
-            }
-            catch (SqlException ex)
-            {
-                throw ex.Number switch
-                {
-                    SqlErrorCodes.NotFound => new StudyNotFoundException(),
-                    _ => new DataStoreException(ex),
-                };
-            }
-        }
-    }
-}
